Introduced BackgroundScript
