+++ source/Background/index.js
@@ -1,14 +1,16 @@
 import extension from 'extensionizer';
-import { init } from './Controller';
+import backgroundScript from './script';
 
 extension.runtime.onInstalled.addListener(async () => {
-  await init();
+  await backgroundScript.init();
+  backgroundScript.exposeHandlers();
   /* eslint-disable-next-line no-console */
   console.log('controller instantiated on install');
 });
 
 extension.runtime.onStartup.addListener(async () => {
-  await init();
+  await backgroundScript.init();
+  backgroundScript.exposeHandlers();
   /* eslint-disable-next-line no-console */
   console.log('controller instantiated on startup');
 });
